Code Restructuring and Responsive Loading
Code Restructuring
- Restructured the app folder to follow some standards I found online
(it was getting ugly).

Responsive Loading
- Added ajaxContent directive which displaying loading images while
loading data, a no content message if the data set is empty, or the data
if everything was ok. For now, this is controlled by a variable in the
scope of a the controller.
- Added ajaxForm directive which disables the form and shows a spinner
while the submit request is being handled. For now, this is controlled
by a variable in the scope of the controller.

Miscellaneous Changes
- Gave the javascript files a handle to the plugin's url so they can
find images and templates (wrm.php).
- Restyled admin table manipulation modals.