Minification
In order to minify and have different versions of the application (dev
and prod), the project now uses NPM (for gulp, specifically). To install
all the required modules, just type "npm install ." while in the
directory of the project and it will install the modules listed in
package.json.

Developer builds are non-minified.
Production builds are minified.

Each of the batch files creates the zip version of the plugin that
WordPress needs in order to upload the plugin (and deletes any temporary
directories used by gulp in the process of doing so).